[qacode] added module for screenshots
+ divided functions on two modules

- screenshots + dummy test
- commons

+ removed property _modules at browser class
